Fix normalizing of sponsor data
When we populate from the DB we have to wait until after the query can be
fetched to noramlize and populate the sponsor data on individual documents to
always have a `display_name`. This was only affecting docs with user sponsors.

Also updated some variables to be pluralized where it makes sense (for `hasMany`
relationships).

Issue #712