Add discussion state to docs, add PATCH route
This includes the first step on #876, which involves using the `FormRequest`
class to validate input coming in for document updates. We provide a `PATCH`
route that will be used to update all attributes on a document, instead of all
the invididual ones we are currently using.

Also adding the document `discussion_state` attribute and define valid values
for this. First step on #874.