Migration: move individual sponsor data to groups
Also created a migration to add `user_id` to groups.
Had to add a couple fillable attributes to `Group`.

Migration does not delete the user_meta data, just in case. On rollback it will
check for the existence of the record and update status according to group
status. Otherwise it will create a new record.

Issue #888