Enable custom favicons during client build
Summary:
The favicon file was not accounted for in the client building process
for Madison 3.0. These changes enable the placement of a custom favicon that
will be picked up during the client build process. The capistrano deployment
process also makes sure to include custom favicons as a linked file. If a custom
favicon is not found, the default file will be used instead.

Test Plan:
* To check the default favicon..
  * Run `make client-build`
  * Make sure `client/app/favicon_default.ico` is copied to `client/build/favicon.ico`
* To check a custom favicon..
  * Put a new favicon file at `client/app/favicon.ico`
  * Run `make client-build`
  * Make sure `client/build/favicon.ico` matches your custom favicon

Reviewers: doshitan

Reviewed By: doshitan

Maniphest Tasks: T16

Differential Revision: https://phabricator.opengovfoundation.org/D2