Update homestead to latest and add "watson/active" to composer.json
Summary:
Vagrant keeps complaining about a newer version of homestead is available and
it's good to keep up, so update it.

Also, somehow "watson/active" never got added to `composer.json`, though it did end up
in the lockfile so things worked. We do depend on it, so it should be listed in
our dependencies.

Test Plan:
- `vagrant box update && vagrant destroy --force && vagrant up`
- Setup the Madison db (`make db-reset` is easiest)
- Ensure your setup still works, run the tests, etc.

Reviewers: sethetter

Reviewed By: sethetter

Differential Revision: https://phabricator.opengovfoundation.org/D182